Viz: apply forces perpendicular to edges
This prevents new branches from sitting *under* their sibling nodes while the
new forces settle out. i.e. it helps resolve to the final shape faster.

It also tries to prevent nodes from crossing edges (although if a new node
spawns on the wrong side of an edge, it will be kept on the wrong side).
If we want to correct existing crossings, we'll need to detect them and
either disable or reverse the forces.