Viz: allow nodes to be dragged around by the mouse
This helps if e.g. the layout gets edges crossed in a way that the
algorithm can't recover from... also it's fun!

The dragged node is effectively removed from force calculations.
Repositioning is easier but less entertaining without the other nodes
reacting to the dragged node's movement (until it is released).
An alternative would be to give it infinite mass (0 inverse mass), or
possibly to remove it from just some of the force calculations.