move gh #132 "Users feel safe about finality of incoming transactions" to sl-crosslink-post-1.0-brainstorms .
Reason: our conversation with Paul Puey of Edge taught us that users are going
to expect *some* kind of finality within seconds or a minute at the most, and
wallets are going to give *some* kind of fast finality indicator to their
users. We can't make finality that fast in the first release of Crosslink, so
let's just make it safe and profitable in the first release and then we can
hopefully make it fast in a future release.